DDS2-2078

+++ app/models/onshore/OnshoreTaxYearLiabilities.scala
@@ -26,7 +26,7 @@ final case class OnshoreTaxYearLiabilities(
   unpaidTax: BigInt,
   niContributions: BigInt,
   interest: BigInt,
-  penaltyRate: Int,
+  penaltyRate: BigDecimal,
   penaltyRateReason: String,
   residentialTaxReduction: Option[Boolean]
 )
